# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vena - Gestão Inteligente de Compras e Obras
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects/obras
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/budget-analysis:
    get:
      operationId: getProjectBudgetAnalysis
      tags: [projects]
      summary: Get budget vs actual analysis for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetAnalysis"

  # MATERIALS / INVENTORY
...
  /price-monitor/products:
    get:
      operationId: listMonitoredProducts
      tags: [price-monitor]
      summary: List monitored products
      responses:
        "200":
          description: Monitored products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonitoredProduct"
    post:
      operationId: createMonitoredProduct
      tags: [price-monitor]
      summary: Add product to price monitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitoredProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoredProduct"
  /price-monitor/products/{id}:
    get:
      operationId: getMonitoredProduct
      tags: [price-monitor]
      summary: Get monitored product with price history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Monitored product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoredProduct"
    delete:
      operationId: deleteMonitoredProduct
      tags: [price-monitor]
      summary: Remove product from monitor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /price-monitor/products/{id}/price-history:
    get:
      operationId: listProductPriceHistory
      tags: [price-monitor]
      summary: Get price history for a monitored product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Price history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PricePoint"
  /price-monitor/supplier-comparison:
    get:
      operationId: getSupplierPriceComparison
      tags: [price-monitor]
      summary: Compare prices across suppliers for a product
      parameters:
        - name: product_name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Supplier comparison
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierPriceEntry"
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, client, status, budget, spent, startDate, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        client:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [planning, active, paused, completed, cancelled]
        budget:
          type: number
        spent:
          type: number
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        manager:
          type: ["string", "null"]
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name, client, budget, startDate]
      properties:
        name:
          type: string
        client:
          type: string
        description:
          type: string
        budget:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        location:
          type: string
        manager:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        client:
          type: string
        description:
          type: string
        status:
          type: string
        budget:
          type: number
        endDate:
          type: string
        location:
          type: string
        manager:
          type: string
    BudgetAnalysis:
      type: object
      required: [projectId, projectName, budget, spent, committed, remaining, percentUsed, byCategory]
      properties:
        projectId:
          type: integer
        projectName:
          type: string
        budget:
          type: number
        spent:
          type: number
        committed:
          type: number
        remaining:
          type: number
        percentUsed:
          type: number
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/BudgetCategoryBreakdown"
    BudgetCategoryBreakdown:
      type: object
      required: [category, budgeted, spent]
      properties:
        category:
          type: string
        budgeted:
          type: number
        spent:
          type: number

    # MATERIALS
...
    MonitoredProduct:
      type: object
      required: [id, name, category, currentPrice, lowestPrice30d, lowestPrice60d, lowestPrice90d, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        url:
          type: ["string", "null"]
        currentPrice:
          type: number
        lowestPrice30d:
          type: number
        lowestPrice60d:
          type: number
        lowestPrice90d:
          type: number
        priceChangePercent:
          type: number
        alertThresholdPercent:
          type: number
        aiInsight:
          type: ["string", "null"]
        createdAt:
          type: string
    MonitoredProductInput:
      type: object
      required: [name, category, currentPrice]
      properties:
        name:
          type: string
        category:
          type: string
        url:
          type: string
        currentPrice:
          type: number
        alertThresholdPercent:
          type: number
    PricePoint:
      type: object
      required: [date, price]
      properties:
        date:
          type: string
        price:
          type: number
        supplier:
          type: ["string", "null"]
    SupplierPriceEntry:
      type: object
      required: [supplier, price, deliveryDays, rating]
      properties:
        supplier:
          type: string
        price:
          type: number
        deliveryDays:
          type: integer
        freight:
          type: ["number", "null"]
        rating:
          type: number
        isRecommended:
          type: boolean
...