# lib/api-zod/src/generated/types/supplierPriceEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vena - Gestão Inteligente de Compras e Obras
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierPriceEntry {
  supplier: string;
  price: number;
  deliveryDays: number;
  /** @nullable */
  freight?: number | null;
  rating: number;
  isRecommended?: boolean;
}